WIP - thinking a little differently about how I want this to work
