+++ app/download_reports.rb
@@ -1,0 +1,5 @@
+require_relative "models/report"
+
+Report.all.each do |report|
+  puts "#{report.year} #{report.url}"
+end
