+++ app/models/report.rb
@@ -1,20 +1,21 @@
-#require "pry"
+require "active_support/core_ext/hash/keys"
+
+require_relative "has_url"
 
 class Report
-  attr_accessor :school_uuid, :class_of, :url, :downloaded_on
+  include HasUrl
+
+  attr_accessor :school_uuid, :class_of, :url, :downloaded_on, :parsed_on
+
+  REPORTS_FILE = File.expand_path("../../../db/reports.csv", __FILE__)
 
   def initialize(options)
     @school_uuid = options[:school_uuid].to_i
     @class_of = options[:class_of].to_i
-    @url = options[:url_source]
+    @url = options[:url]
     @downloaded_on = options[:downloaded_on]
+    @parsed_on = options[:parsed_on]
   end
-
-  #
-  # READ INFO FROM CSV FILE
-  #
-
-  REPORTS_FILE = File.expand_path("../../../db/reports.csv", __FILE__)
 
   def self.all
     @all ||= CSV.read(REPORTS_FILE, headers: true).map{|row| self.new(row.to_h.symbolize_keys) }
@@ -26,5 +27,17 @@ class Report
 
   def year
     class_of
+  end
+
+  def file_format
+    #code
+  end
+
+  def downloaded?
+    #code
+  end
+
+  def parsed?
+    #code
   end
 end
