+++ app/models/school.rb
@@ -1,24 +1,23 @@
-require "httparty"
-require "domainatrix"
 require "active_support/core_ext/hash/keys"
 require "fileutils"
+require "httparty"
+
+require_relative "has_url"
 
 class School
+  include HasUrl
+
   attr_accessor :uuid, :alt_name, :name, :year_founded, :url
+
+  SCHOOLS_FILE = File.expand_path("../../../db/schools.csv", __FILE__)
 
   def initialize(options)
     @uuid = options[:uuid].to_i
     @name = options[:name]
-    @alt_name = options[:name]
+    @alt_name = options[:alt_name]
     @year_founded = options[:year_founded].to_i
     @url = options[:url]
   end
-
-  #
-  # READ INFO FROM CSV FILE
-  #
-
-  SCHOOLS_FILE = File.expand_path("../../../db/schools.csv", __FILE__)
 
   def self.all
     @all ||= CSV.read(SCHOOLS_FILE, headers: true).map{|row| self.new(row.to_h.symbolize_keys) }
@@ -32,24 +31,12 @@ class School
     all_domains.select{|domain| all_domains.count(domain) > 1}.uniq #> ["psu", "widener"]
   end
 
-  def parsed_url
-    Domainatrix.parse(url)
-  end
-
-  def domain
-    parsed_url.domain
-  end
-
-  def subdomain
-    parsed_url.subdomain
-  end
-
-  def host
-    parsed_url.host
-  end
-
   def short_name
     self.class.duplicate_domains.include?(domain) ? subdomain : domain
+  end
+
+  def title_name
+    "#{name.upcase} (#{short_name.upcase})"
   end
 
   #
